Class Seperation
- Split menu and card model off into their own files
- Built a start/welcome screen
